# static/openapi.yaml
openapi: 3.0.3
info:
  title: GitHub Stars Webhook Limiter API
  version: 2.4.2
  description: 'A Flask application that listens for GitHub webhook events (star/watch),
    validates them,

    and sends notifications to Discord webhooks if a user performs an action for the
    first time.


    ## Features

    - Validates GitHub webhook secrets using HMAC

    - Prevents duplicate notifications for the same user/repo pair

    - Sends Discord notifications for new stars/watches

    - RESTful API for repository management

    - Admin panel for API key management

    - Rate limiting per API key


    ## API Testing

    - **Interactive Testing**: Visit [/try](/try) to test API endpoints with your
    API key using Swagger UI

    - **Documentation**: Visit [/docs](/docs) for comprehensive API documentation

    - **OpenAPI Spec**: Access the raw specification at [/spec](/spec)


    ## Links

    - **Repository**: [GitHub](https://github.com/Serpensin/GitHub-Stars-Webhook-Limiter)

    - **Discord**: [Join our server](https://url.serpensin.com/discord)

    '
  contact:
    name: Serpensin
    url: https://serpensin.com
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
- url: /
  description: Current server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "API key for accessing public endpoints and admin endpoints. \n\
        \n**For Public API endpoints**: Get your API key from the Admin Panel.\n\n\
        **For Admin API endpoints**: Use an admin API key (created with `is_admin_key:\
        \ true`).\n\n**Testing**: Click the \"Authorize\" button above to enter your\
        \ API key and test endpoints interactively.\n"
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: 'Admin session cookie (automatically set when accessing the admin
        panel at /admin).


        Note: When using Swagger UI, admin endpoints can be accessed using an admin
        API key in the bearerAuth scheme.

        '
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required:
      - error
...
paths:
...
  /api/stats:
    get:
      tags:
      - Public API
      summary: Get system statistics
      description: 'Retrieve comprehensive statistics about the system including:

        - Current counts of repositories and API keys

        - Total events received (unique and duplicates)

        - Deletion statistics by reason

        - Top users by valid and invalid events

        - Discord message statistics


        **Requires Authentication:** API key with `stats` permission or admin session.


        **Note:** A public version of this endpoint is available at `/stats` (without
        `/api` prefix)

        that doesn''t require authentication when accessed from a browser.

        '
      security:
      - bearerAuth: []
      - sessionAuth: []
      responses:
        '200':
          description: Statistics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  totals:
                    type: object
                    properties:
                      repositories_current:
                        type: integer
                        description: Number of repositories currently registered
                        example: 42
                      api_keys_current:
                        type: integer
                        description: Total number of API keys currently in the system
                        example: 15
                      api_keys_active:
                        type: integer
                        description: Number of API keys currently active (is_active
                          = true)
                        example: 12
                      api_keys_admin:
                        type: integer
                        description: Number of admin API keys (with full access)
                        example: 2
                      repositories_ever_added:
                        type: integer
                        description: Total number of repositories ever added
                        example: 67
                      api_keys_ever_created:
                        type: integer
                        description: Total number of API keys ever created
                        example: 23
                      discord_messages_sent:
                        type: integer
                        description: Total Discord notifications sent
                        example: 1234
                      events_received:
                        type: integer
                        description: Total webhook events received
                        example: 5678
                      unique_events:
                        type: integer
                        description: Number of unique events (first-time user/repo
                          combinations)
                        example: 890
                      duplicate_events:
                        type: integer
                        description: Number of duplicate events (already seen user/repo
                          combinations)
                        example: 4788
                  deletions:
                    type: object
                    properties:
                      repos_inactive_360_days:
                        type: integer
                        description: Repositories deleted due to 360 days of inactivity
                        example: 12
                      repos_github_deleted:
                        type: integer
                        description: Repositories deleted because they were removed
                          from GitHub
                        example: 5
                      repos_webhook_invalid:
                        type: integer
                        description: Repositories deleted due to invalid webhook configuration
                        example: 3
                      api_keys_inactive_360_days:
                        type: integer
                        description: API keys deleted due to 360 days of inactivity
                        example: 7
                  top_users:
                    type: object
                    properties:
                      valid_events:
                        type: array
                        description: Top 10 users by number of valid events
                        items:
                          type: object
                          properties:
                            user_id:
                              type: integer
                              description: GitHub user ID
                              example: 12345
                            username:
                              type: string
                              description: GitHub username
                              example: octocat
                            count:
                              type: integer
                              description: Number of valid events
                              example: 42
                      invalid_events:
                        type: array
                        description: Top 10 users by number of invalid events
                        items:
                          type: object
                          properties:
                            user_id:
                              type: integer
                              description: GitHub user ID
                              example: 67890
                            username:
                              type: string
                              description: GitHub username
                              example: hacker
                            count:
                              type: integer
                              description: Number of invalid events
                              example: 15
                  timestamp:
                    type: integer
                    description: Unix timestamp when the statistics were generated
                    example: 1699000000
              example:
                totals:
                  repositories_current: 42
                  api_keys_current: 15
                  api_keys_active: 12
                  api_keys_admin: 2
                  repositories_ever_added: 67
                  api_keys_ever_created: 23
                  discord_messages_sent: 1234
                  events_received: 5678
                  unique_events: 890
                  duplicate_events: 4788
                deletions:
                  repos_inactive_360_days: 12
                  repos_github_deleted: 5
                  repos_webhook_invalid: 3
                  api_keys_inactive_360_days: 7
                top_users:
                  valid_events:
                  - user_id: 12345
                    username: octocat
                    count: 42
                  invalid_events:
                  - user_id: 67890
                    username: hacker
                    count: 15
                timestamp: 1699000000
        '401':
          description: Unauthorized - Missing or invalid API key/session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Unauthorized
        '403':
          description: Forbidden - API key lacks 'stats' permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: API key does not have the required 'stats' permission
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Failed to retrieve statistics
...
  /webhook:
    post:
      tags:
      - Webhooks
      summary: GitHub webhook receiver
      description: "Receives GitHub webhook events for star and watch actions.\n\n\
        ### Setup Instructions:\n1. Go to your GitHub repository Settings \u2192 Webhooks\n\
        2. Click \"Add webhook\"\n3. Set Payload URL: `https://your-domain.com/webhook`\n\
        4. Content type: `application/json`\n5. Secret: Use the secret from `/api/generate-secret`\n\
        6. Select events: **Stars** and **Watches**\n7. Click \"Add webhook\"\n\n\
        The application will verify the signature and send Discord notifications\n\
        for first-time interactions only.\n"
      parameters:
      - name: X-GitHub-Event
        in: header
        required: true
        schema:
          type: string
          enum:
          - star
          - watch
        description: GitHub event type
      - name: X-Hub-Signature-256
        in: header
        required: true
        schema:
          type: string
        description: HMAC SHA-256 signature of the payload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: GitHub webhook payload (format varies by event type)
      responses:
        '200':
          description: Event processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Notification sent
        '400':
          description: Bad request - Invalid payload or missing headers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - Invalid signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Repository not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'